Move Codelist domain class to its own package
This mirrors the structure in the MDC, and more classes will be added in
this package.
