+++ src/main/java/eu/europa/ted/eforms/sdk/domain/codelist/Codelist.java
@@ -1,4 +1,4 @@
-package eu.europa.ted.eforms.sdk.domain;
+package eu.europa.ted.eforms.sdk.domain.codelist;
 
 import java.io.Serializable;
 import java.util.Set;
